requirements.txt
requests
selenium
flask
uvicorn[standard]
python-dotenv
asgiref
